+++ Duckling/Ranking/Classifiers/ES_CO.hs
@@ -48,11 +48,11 @@ classifiers
                                likelihoods = HashMap.fromList [], n = 0}}),
        ("integer (numeric)",
         Classifier{okData =
-                     ClassData{prior = -0.5465437063680699, unseen = -4.219507705176107,
-                               likelihoods = HashMap.fromList [("", 0.0)], n = 66},
+                     ClassData{prior = -0.5930637220029626, unseen = -4.174387269895637,
+                               likelihoods = HashMap.fromList [("", 0.0)], n = 63},
                    koData =
-                     ClassData{prior = -0.8649974374866046, unseen = -3.912023005428146,
-                               likelihoods = HashMap.fromList [("", 0.0)], n = 48}}),
+                     ClassData{prior = -0.8043728156701697, unseen = -3.970291913552122,
+                               likelihoods = HashMap.fromList [("", 0.0)], n = 51}}),
        ("the day before yesterday",
         Classifier{okData =
                      ClassData{prior = 0.0, unseen = -1.3862943611198906,
@@ -97,63 +97,64 @@ classifiers
                                likelihoods = HashMap.fromList [], n = 0}}),
        ("intersect by `de`",
         Classifier{okData =
-                     ClassData{prior = -0.1416505170630268, unseen = -4.875197323201151,
+                     ClassData{prior = -0.19290366612449156,
+                               unseen = -4.890349128221754,
                                likelihoods =
                                  HashMap.fromList
-                                   [("Miercoleseste|en un <cycle>", -4.174387269895637),
+                                   [("Miercoleseste|en un <cycle>", -4.189654742026425),
                                     ("<day-of-month> de <named-month>year (value by adding three composing numbers together)",
-                                     -4.174387269895637),
+                                     -4.189654742026425),
                                     ("el <day-of-month> (non ordinal)intersect by `de`",
-                                     -3.7689221617874726),
-                                    ("daymonth", -1.8230120127321594),
-                                    ("monthyear", -3.7689221617874726),
-                                    ("el <day-of-month> (non ordinal)Abril", -4.174387269895637),
-                                    ("Marteseste|en un <cycle>", -4.174387269895637),
-                                    ("Miercolesel <cycle> (proximo|que viene)", -4.174387269895637),
-                                    ("el <time>Marzo", -3.7689221617874726),
+                                     -3.784189633918261),
+                                    ("daymonth", -1.8382794848629478),
+                                    ("monthyear", -3.784189633918261),
+                                    ("el <day-of-month> (non ordinal)Abril", -4.189654742026425),
+                                    ("Marteseste|en un <cycle>", -4.189654742026425),
+                                    ("Miercolesel <cycle> (proximo|que viene)", -4.189654742026425),
+                                    ("el <time>Marzo", -3.784189633918261),
                                     ("el <day-of-month> (non ordinal)Septiembre",
-                                     -4.174387269895637),
-                                    ("Miercolesel proximo <cycle> ", -4.174387269895637),
-                                    ("Luneseste|en un <cycle>", -4.174387269895637),
-                                    ("dayyear", -2.159484249353372),
-                                    ("dd-dd <month>(interval)year", -4.174387269895637),
-                                    ("el <time>el <cycle> (proximo|que viene)", -4.174387269895637),
-                                    ("el <day-of-month> (non ordinal)Julio", -4.174387269895637),
-                                    ("el <time>year", -3.7689221617874726),
-                                    ("Septiembreyear", -4.174387269895637),
-                                    ("el <time>la <cycle> pasado", -4.174387269895637),
-                                    ("day of month (1st)Marzo", -3.258096538021482),
-                                    ("el <day-of-month> de <named-month>year", -3.7689221617874726),
-                                    ("el <time>este|en un <cycle>", -3.7689221617874726),
-                                    ("Mayoyear", -4.174387269895637),
-                                    ("<day-of-month> de <named-month>year", -3.7689221617874726),
-                                    ("two time tokens separated by \",\"year", -3.481240089335692),
-                                    ("el <day-of-month> (non ordinal)Febrero", -4.174387269895637),
+                                     -4.189654742026425),
+                                    ("Miercolesel proximo <cycle> ", -4.189654742026425),
+                                    ("Luneseste|en un <cycle>", -4.189654742026425),
+                                    ("dayyear", -2.1102132003465894),
+                                    ("dd-dd <month>(interval)year", -4.189654742026425),
+                                    ("el <time>el <cycle> (proximo|que viene)", -4.189654742026425),
+                                    ("el <day-of-month> (non ordinal)Julio", -4.189654742026425),
+                                    ("el <time>year", -3.784189633918261),
+                                    ("Septiembreyear", -4.189654742026425),
+                                    ("el <time>la <cycle> pasado", -4.189654742026425),
+                                    ("day of month (1st)Marzo", -3.2733640101522705),
+                                    ("el <day-of-month> de <named-month>year", -3.784189633918261),
+                                    ("el <time>este|en un <cycle>", -3.784189633918261),
+                                    ("Mayoyear", -4.189654742026425),
+                                    ("<day-of-month> de <named-month>year", -3.4965075614664802),
+                                    ("two time tokens separated by \",\"year", -3.4965075614664802),
+                                    ("el <day-of-month> (non ordinal)Febrero", -4.189654742026425),
                                     ("two time tokens separated by \",\"intersect by `de`",
-                                     -4.174387269895637),
-                                    ("dayweek", -2.469639177657212),
-                                    ("intersect by `de`year", -3.481240089335692),
-                                    ("el <day-of-month> (non ordinal)Marzo", -3.481240089335692),
-                                    ("Domingola <cycle> pasado", -4.174387269895637),
+                                     -4.189654742026425),
+                                    ("dayweek", -2.4849066497880004),
+                                    ("intersect by `de`year", -3.4965075614664802),
+                                    ("el <day-of-month> (non ordinal)Marzo", -3.4965075614664802),
+                                    ("Domingola <cycle> pasado", -4.189654742026425),
                                     ("two time tokens separated by \",\"Septiembre",
-                                     -4.174387269895637),
-                                    ("el <day-of-month> (non ordinal)Mayo", -3.481240089335692)],
-                               n = 46},
+                                     -4.189654742026425),
+                                    ("el <day-of-month> (non ordinal)Mayo", -3.4965075614664802)],
+                               n = 47},
                    koData =
-                     ClassData{prior = -2.0243817644968085, unseen = -3.970291913552122,
+                     ClassData{prior = -1.7404661748405046, unseen = -4.07753744390572,
                                likelihoods =
                                  HashMap.fromList
                                    [("Mayoyear (value by adding three composing numbers together)",
-                                     -3.258096538021482),
-                                    ("monthyear", -2.5649493574615367),
-                                    ("Julioyear", -3.258096538021482),
-                                    ("hourmonth", -2.8526314299133175),
-                                    ("a las <time-of-day>Julio", -3.258096538021482),
-                                    ("dayyear", -2.8526314299133175),
-                                    ("Mayoyear", -3.258096538021482),
-                                    ("<day-of-month> de <named-month>year", -2.8526314299133175),
-                                    ("a las <time-of-day>Enero", -3.258096538021482)],
-                               n = 7}}),
+                                     -3.367295829986474),
+                                    ("monthyear", -2.268683541318364),
+                                    ("Julioyear", -3.367295829986474),
+                                    ("hourmonth", -2.9618307218783095),
+                                    ("a las <time-of-day>Julio", -3.367295829986474),
+                                    ("dayyear", -2.6741486494265287),
+                                    ("Mayoyear", -2.6741486494265287),
+                                    ("<day-of-month> de <named-month>year", -2.6741486494265287),
+                                    ("a las <time-of-day>Enero", -3.367295829986474)],
+                               n = 10}}),
        ("n pasados <cycle>",
         Classifier{okData =
                      ClassData{prior = 0.0, unseen = -2.1972245773362196,
@@ -442,26 +443,26 @@ classifiers
                                likelihoods = HashMap.fromList [], n = 0}}),
        ("a las <time-of-day>",
         Classifier{okData =
-                     ClassData{prior = -5.770831762064673e-2,
-                               unseen = -5.365976015021851,
+                     ClassData{prior = -7.320340402329494e-2,
+                               unseen = -5.135798437050262,
                                likelihoods =
                                  HashMap.fromList
-                                   [("<hour-of-day> and half", -3.9749978045895347),
-                                    ("<time-of-day> horas", -4.66814498514948),
-                                    ("<hour-of-day> and quarter", -3.751854253275325),
-                                    ("time-of-day (latent)", -1.6477200990051173),
-                                    ("<hour-of-day> and <relative minutes>", -3.56953269648137),
-                                    ("<time-of-day> am|pm", -2.1832383353614793),
+                                   [("<hour-of-day> and half", -3.7436043538031827),
+                                    ("<time-of-day> horas", -4.436751534363128),
+                                    ("<hour-of-day> and quarter", -3.520460802488973),
+                                    ("time-of-day (latent)", -1.4163266482187657),
+                                    ("<hour-of-day> and <relative minutes>", -3.3381392456950185),
+                                    ("<time-of-day> am|pm", -2.5649493574615367),
                                     ("<hour-of-day> minus <integer> (as relative minutes)",
-                                     -4.66814498514948),
+                                     -4.436751534363128),
                                     ("<hour-of-day> minus quarter (as relative minutes)",
-                                     -4.2626798770413155),
-                                    ("hour", -1.6000920500158626), ("minute", -1.2669476034873244),
+                                     -4.0312864262549635),
+                                    ("hour", -1.3686985992295109), ("minute", -1.4923125551966876),
                                     ("<hour-of-day> <integer> (as relative minutes)",
-                                     -2.2257979497802753)],
-                               n = 101},
+                                     -2.644992065135073)],
+                               n = 79},
                    koData =
-                     ClassData{prior = -2.8810693652338513,
+                     ClassData{prior = -2.6508917872622613,
                                unseen = -3.1780538303479458,
                                likelihoods =
                                  HashMap.fromList
@@ -642,6 +643,23 @@ classifiers
                    koData =
                      ClassData{prior = -infinity, unseen = -0.6931471805599453,
                                likelihoods = HashMap.fromList [], n = 0}}),
+       ("<thousands> 0..999",
+        Classifier{okData =
+                     ClassData{prior = -0.6931471805599453,
+                               unseen = -1.6094379124341003,
+                               likelihoods =
+                                 HashMap.fromList
+                                   [("1K or 1M in multipliable form<hundreds> 0..99",
+                                     -0.2876820724517809)],
+                               n = 2},
+                   koData =
+                     ClassData{prior = -0.6931471805599453,
+                               unseen = -1.6094379124341003,
+                               likelihoods =
+                                 HashMap.fromList
+                                   [("1K or 1M in multipliable formbig number 100 to 1K",
+                                     -0.2876820724517809)],
+                               n = 2}}),
        ("now",
         Classifier{okData =
                      ClassData{prior = 0.0, unseen = -1.791759469228055,
@@ -676,13 +694,12 @@ classifiers
                                likelihoods = HashMap.fromList [], n = 0}}),
        ("number (0..15)",
         Classifier{okData =
-                     ClassData{prior = -2.381064869371856e-2,
-                               unseen = -4.442651256490317,
-                               likelihoods = HashMap.fromList [("", 0.0)], n = 83},
+                     ClassData{prior = -0.13858616328614667,
+                               unseen = -4.330733340286331,
+                               likelihoods = HashMap.fromList [("", 0.0)], n = 74},
                    koData =
-                     ClassData{prior = -3.7495040759303713,
-                               unseen = -1.3862943611198906,
-                               likelihoods = HashMap.fromList [("", 0.0)], n = 2}}),
+                     ClassData{prior = -2.044755983691946, unseen = -2.5649493574615367,
+                               likelihoods = HashMap.fromList [("", 0.0)], n = 11}}),
        ("D\237a Internacional de las Cooperativas",
         Classifier{okData =
                      ClassData{prior = 0.0, unseen = -1.0986122886681098,
@@ -786,11 +803,11 @@ classifiers
                                likelihoods = HashMap.fromList [("", 0.0)], n = 6}}),
        ("1K or 1M in multipliable form",
         Classifier{okData =
-                     ClassData{prior = -infinity, unseen = -0.6931471805599453,
-                               likelihoods = HashMap.fromList [], n = 0},
-                   koData =
                      ClassData{prior = 0.0, unseen = -1.3862943611198906,
-                               likelihoods = HashMap.fromList [("", 0.0)], n = 2}}),
+                               likelihoods = HashMap.fromList [("", 0.0)], n = 2},
+                   koData =
+                     ClassData{prior = -infinity, unseen = -0.6931471805599453,
+                               likelihoods = HashMap.fromList [], n = 0}}),
        ("afternoon",
         Classifier{okData =
                      ClassData{prior = -0.2876820724517809,
@@ -802,21 +819,20 @@ classifiers
                                likelihoods = HashMap.fromList [("", 0.0)], n = 2}}),
        ("time-of-day (latent)",
         Classifier{okData =
-                     ClassData{prior = -0.7884573603642702, unseen = -3.912023005428146,
+                     ClassData{prior = -0.6592456288842639, unseen = -3.891820298110627,
                                likelihoods =
                                  HashMap.fromList
-                                   [("integer (numeric)", -1.3268709406490897),
-                                    ("number (0..15)", -0.36545977349446523)],
+                                   [("integer (numeric)", -1.3062516534463542),
+                                    ("number (0..15)", -0.3448404862917295)],
                                n = 45},
                    koData =
-                     ClassData{prior = -0.6061358035703156, unseen = -4.07753744390572,
+                     ClassData{prior = -0.7282385003712154, unseen = -3.828641396489095,
                                likelihoods =
                                  HashMap.fromList
-                                   [("integer (numeric)", -1.7578579175523736),
-                                    ("number (0..15)", -0.9694005571881036),
-                                    ("<hundreds> 0..99", -1.4954936530848826),
-                                    ("integer (0-9) with two digits", -1.4954936530848826)],
-                               n = 54}}),
+                                   [("integer (numeric)", -1.5040773967762742),
+                                    ("number (0..15)", -0.7156200364120039),
+                                    ("integer (0-9) with two digits", -1.241713132308783)],
+                               n = 42}}),
        ("<hour-of-day> and <relative minutes>",
         Classifier{okData =
                      ClassData{prior = -0.3364722366212129, unseen = -3.332204510175204,
@@ -842,19 +858,22 @@ classifiers
                                n = 4}}),
        ("year",
         Classifier{okData =
-                     ClassData{prior = -0.8754687373538999,
-                               unseen = -2.0794415416798357,
-                               likelihoods =
-                                 HashMap.fromList [("integer (numeric)", -0.15415067982725836)],
-                               n = 5},
-                   koData =
-                     ClassData{prior = -0.5389965007326869,
-                               unseen = -2.3025850929940455,
+                     ClassData{prior = -0.8266785731844679,
+                               unseen = -2.3978952727983707,
                                likelihoods =
                                  HashMap.fromList
-                                   [("integer (numeric)", -0.40546510810816444),
-                                    ("1K or 1M in multipliable form", -1.0986122886681098)],
-                               n = 7}}),
+                                   [("integer (numeric)", -0.5108256237659907),
+                                    ("<thousands> 0..999", -1.2039728043259361)],
+                               n = 7},
+                   koData =
+                     ClassData{prior = -0.5753641449035618,
+                               unseen = -2.5649493574615367,
+                               likelihoods =
+                                 HashMap.fromList
+                                   [("integer (numeric)", -0.6931471805599453),
+                                    ("<thousands> 0..999", -1.3862943611198906),
+                                    ("1K or 1M in multipliable form", -1.3862943611198906)],
+                               n = 9}}),
        ("en <duration>",
         Classifier{okData =
                      ClassData{prior = 0.0, unseen = -3.6635616461296463,
@@ -950,25 +969,24 @@ classifiers
                                likelihoods = HashMap.fromList [], n = 0}}),
        ("<time-of-day> am|pm",
         Classifier{okData =
-                     ClassData{prior = -0.5306282510621704, unseen = -4.983606621708336,
+                     ClassData{prior = -0.7069405026922811, unseen = -4.356708826689592,
                                likelihoods =
                                  HashMap.fromList
-                                   [("a las <time-of-day>", -1.798679912072629),
-                                    ("time-of-day (latent)", -4.283586561860629),
-                                    ("hour", -3.8781214537524646), ("minute", -0.742627237823315),
+                                   [("a las <time-of-day>", -1.7788560643921472),
+                                    ("time-of-day (latent)", -3.6506582412937383),
+                                    ("hour", -3.245193133185574), ("minute", -0.7884573603642702),
                                     ("<hour-of-day> <integer> (as relative minutes)",
-                                     -1.1265861407105158)],
-                               n = 70},
+                                     -1.1657515915057381)],
+                               n = 36},
                    koData =
-                     ClassData{prior = -0.8873031950009028,
-                               unseen = -4.6443908991413725,
+                     ClassData{prior = -0.6795415285041666, unseen = -4.382026634673881,
                                likelihoods =
                                  HashMap.fromList
-                                   [("time-of-day (latent)", -0.99714282850325),
-                                    ("hour", -0.99714282850325), ("minute", -2.069779630768099),
+                                   [("time-of-day (latent)", -1.1113513144455396),
+                                    ("hour", -1.1113513144455396), ("minute", -1.8044984950054848),
                                     ("<hour-of-day> <integer> (as relative minutes)",
-                                     -2.069779630768099)],
-                               n = 49}}),
+                                     -1.8044984950054848)],
+                               n = 37}}),
        ("n proximas <cycle>",
         Classifier{okData =
                      ClassData{prior = 0.0, unseen = -2.5649493574615367,
@@ -1067,19 +1085,13 @@ classifiers
                                likelihoods = HashMap.fromList [], n = 0}}),
        ("<hundreds> 0..99",
         Classifier{okData =
-                     ClassData{prior = -0.15415067982725836,
-                               unseen = -2.772588722239781,
+                     ClassData{prior = 0.0, unseen = -1.3862943611198906,
                                likelihoods =
-                                 HashMap.fromList
-                                   [("number (0..15)number (0..15)", -0.40546510810816444),
-                                    ("integer (numeric)integer (numeric)", -1.3217558399823195)],
-                               n = 12},
+                                 HashMap.fromList [("big number 100 to 1Knumber (20..90)", 0.0)],
+                               n = 2},
                    koData =
-                     ClassData{prior = -1.9459101490553135, unseen = -1.791759469228055,
-                               likelihoods =
-                                 HashMap.fromList
-                                   [("big number 100 to 1Knumber (20..90)", -0.5108256237659907)],
-                               n = 2}}),
+                     ClassData{prior = -infinity, unseen = -0.6931471805599453,
+                               likelihoods = HashMap.fromList [], n = 0}}),
        ("<hour-of-day> minus <integer> (as relative minutes)",
         Classifier{okData =
                      ClassData{prior = 0.0, unseen = -2.0794415416798357,
@@ -1390,27 +1402,25 @@ classifiers
                                likelihoods = HashMap.fromList [], n = 0}}),
        ("<hour-of-day> <integer> (as relative minutes)",
         Classifier{okData =
-                     ClassData{prior = -0.22738984219566089,
-                               unseen = -4.624972813284271,
+                     ClassData{prior = -0.40546510810816444,
+                               unseen = -3.9889840465642745,
                                likelihoods =
                                  HashMap.fromList
                                    [("a las <time-of-day>integer (0-9) with two digits",
-                                     -2.1302138670532593),
-                                    ("time-of-day (latent)<hundreds> 0..99", -2.0501711593797225),
+                                     -1.4853852637641216),
                                     ("time-of-day (latent)integer (0-9) with two digits",
-                                     -2.0501711593797225),
-                                    ("time-of-day (latent)number (0..15)", -3.9219733362813143),
-                                    ("a las <time-of-day><hundreds> 0..99", -2.1302138670532593),
-                                    ("hour", -0.7439195059333685)],
-                               n = 47},
+                                     -1.405342556090585),
+                                    ("time-of-day (latent)number (0..15)", -3.2771447329921766),
+                                    ("hour", -0.7514160886839211)],
+                               n = 24},
                    koData =
-                     ClassData{prior = -1.5926307941177191,
-                               unseen = -3.4657359027997265,
+                     ClassData{prior = -1.0986122886681098,
+                               unseen = -3.4011973816621555,
                                likelihoods =
                                  HashMap.fromList
-                                   [("time-of-day (latent)integer (numeric)", -2.0476928433652555),
-                                    ("time-of-day (latent)number (0..15)", -1.1314021114911006),
-                                    ("hour", -0.8690378470236094)],
+                                   [("time-of-day (latent)integer (numeric)", -1.9810014688665833),
+                                    ("time-of-day (latent)number (0..15)", -1.0647107369924282),
+                                    ("hour", -0.8023464725249373)],
                                n = 12}}),
        ("integer (0-9) with two digits",
         Classifier{okData =